requirements.txt
python-telegram-bot==20.7
pillow==9.2.0
tensorflow==2.15.0
numpy
chess
aiohttp==3.9.3
nest-asyncio==1.6.0
requests==2.31.0
